Replace Claude with free local AI answers (Ollama or extractive)
Remove the Anthropic dependency. Ask AI now uses free, local models:

- Ollama (preferred): when a local Ollama server has OLLAMA_MODEL
  (default llama3.2) pulled, sources are sent to /api/chat and the answer
  streams back with [n] citations. Detected automatically; falls back if
  the server is unreachable.
- Extractive (zero setup): splits the retrieved sources into prose
  sentences, embeds them with the sentence-transformers model already used
  for search, and returns the best-matching sentences with citations.

Also store file mtimes as integer nanoseconds so incremental re-indexing
compares exactly (float mtimes did not always round-trip through SQLite
JSON, making unchanged files look modified).

+++ vector_db.py
@@ -165,7 +165,7 @@ class VectorDB:
         self.conn.commit()
         return cur.rowcount
 
-    def indexed_sources(self, root: str) -> Dict[str, float]:
+    def indexed_sources(self, root: str) -> Dict[str, int]:
         """Return {source_path: mtime} for files indexed from the given root folder."""
         rows = self.conn.execute(
             "SELECT DISTINCT json_extract(metadata, '$.source') AS source, "
